Create README
JSON beutifier is an online tool. it will help to edit JSON formatter online. JSON is used to represent structured data, such as objects, arrays, numbers, strings, booleans, and null values. It provides a way to organize and store data in a format that can be easily understood by both humans and machines.